Extract functions to init_appbuilder module
